Realizo algunas optimizaciones

+++ cpp_cpu_funciones.h
@@ -1,8 +1,8 @@
 
-void f(double* y, double* dydt, double alpha, int N);
-double f_maxwell();
-void condiciones_iniciales(double* y0, int N);
-void distancia_al_origen(double* r_vec, double* d_vec, int N);
-void rebote_blando(double rx, double ry, double vx, double vy, double *result);
-void metodoVerlet(double* yold, double t, double dt, int N, double* ynew, double alpha);
-void avanzo_dt(double* y, double* ynew, double t, double dt, int N, double alpha);
+void f(float* y, float* dydt, float alpha, int N);
+float f_maxwell();
+void condiciones_iniciales(float* y0, int N);
+void distancia_al_origen(float* r_vec, float* d_vec, int N);
+void rebote_blando(float rx, float ry, float vx, float vy, float *result);
+void metodoVerlet(float* yold, float t, float dt, int N, float* ynew, float alpha);
+void avanzo_dt(float* y, float* ynew, float t, float dt, int N, float alpha);
